LLM-PDF1/requirements.txt
streamlit
pandas
fireworks-ai
chromadb
sentence-transformers
bert-score
pypdf # Assuming this is used by your parse_ingest.py
Pillow # Often needed for image handling with Streamlit and some models
torch # or tensorflow, for bert-score